refactor(web): extract the pane shell into PaneHost.svelte (TASK-2170)
Increment 2/2 of PLAN-2154 Phase 2 (Architecture E, bullet 1). Pull the
right-docked detail-pane <aside> SHELL out of the 4805-line collection
route into a new reusable web/src/lib/components/collections/PaneHost.svelte
so both the collection page and (in TASK-2174) the full-page item host
mount ONE shell.

Moved into PaneHost: the .pane-divider + <aside class="item-pane"> markup
and its {#if paneMintForRoute}<ItemDetail embedded .../> render; the resize
machinery (paneEl/paneWidth/storedPaneWidth/measuredPaneWidth/ariaMin/ariaMax
state, PANE_WIDTH_* consts, paneBounds/clampPaneWidth/readStoredPaneWidth/
setPaneWidth/fittedPaneWidth/refitPaneWidth/currentPaneWidth, the
ResizeObserver effect, the divider pointer/keyboard handlers); the focus
machinery (focusPaneRegion, the mobile focus-into effect, the mobile
focus-trap effect, the desktop focusin backstop); and the .item-pane /
.pane-divider / mobile @media CSS. The mid-drag body-chrome cleanup that the
route ran from an openItemRef→null effect is now a PaneHost onDestroy — the
shell's unmount IS that transition.

Stays in the route (unchanged): the {#if openItemRef}<PaneHost/>{/if} gate,
paneMint (paneMintRef/paneMintPathname/paneMintForRoute + createPaneMintSettle
+ its afterNavigate + onDestroy cancel — its always-mounted lifecycle is
load-bearing for TASK-2166 popstate coalescing, so the route computes
paneMintForRoute and passes it as a PROP), createPaneController,
schedulePaneFollow/cancelPaneFollow, focusedItemId + pane-snap,
returnFocusToList/resolvePaneReturnTarget, dismissPaneOnBackgroundClick,
handlePageKeydown.

Seam: PaneHost exposes focusPaneRegion via a Svelte 5 instance export; the
route binds bind:this={paneHostEl} and rewires both the controller's
focusPaneRegion dep and the list→pane Tab bridge to
paneHostEl?.focusPaneRegion(). Behavior-preserving refactor — the load-bearing
no-{#key} invariant is intact (an A→B switch stays a ref PROP UPDATE reusing
the one mounted ItemDetail instance + its single collab provider + SSE sub;
the only mount/unmount is the outer {#if openItemRef}).

Gates: svelte-check 0 errors; vitest 435 passed; pane e2e 49 passed / 49
skipped / 0 failed (the pre-extraction baseline).
